# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JCEmpleos API",
        "description": "Documentación oficial del API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/perfil/experiencia": {
            "get": {
                "tags": [
                    "Perfil - Experiencia"
                ],
                "summary": "Listar experiencias laborales del usuario",
                "operationId": "72883b70f11ef346b2fe4707981053b5",
                "responses": {
                    "200": {
                        "description": "Listado obtenido correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Perfil - Experiencia"
                ],
                "summary": "Agregar experiencia laboral",
                "operationId": "104987c433a99c619f32827e80f99013",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "empresa",
                                    "cargo"
                                ],
                                "properties": {
                                    "empresa": {
                                        "type": "string",
                                        "example": "Empresa XYZ"
                                    },
                                    "pais_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cargo": {
                                        "type": "string",
                                        "example": "Desarrollador Backend"
                                    },
                                    "fecha_desde": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-01-01"
                                    },
                                    "fecha_hasta": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Desarrollo de APIs en Laravel."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Experiencia agregada correctamente"
                    },
                    "422": {
                        "description": "Error al agregar experiencia"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/perfil/experiencia/{id}": {
            "put": {
                "tags": [
                    "Perfil - Experiencia"
                ],
                "summary": "Actualizar experiencia laboral",
                "operationId": "6a876e76b9a32ea8781fa70926a13858",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "empresa": {
                                        "type": "string"
                                    },
                                    "pais_id": {
                                        "type": "integer"
                                    },
                                    "cargo": {
                                        "type": "string"
                                    },
                                    "fecha_desde": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "fecha_hasta": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Experiencia actualizada correctamente"
                    },
                    "422": {
                        "description": "Error al actualizar experiencia"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Perfil - Experiencia"
                ],
                "summary": "Eliminar experiencia laboral",
                "operationId": "8ce6d989c59126df92d04324d59027c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Experiencia eliminada correctamente"
                    },
                    "422": {
                        "description": "Error al eliminar experiencia"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}